HV: add pcpu id check before send IPI
to avoid send IPI to self, also improve the related code:
1. get_cpu_id is uint16_t now
2. MISRA-C requirement. like add {}

Acked-by: Eddie Dong <eddie.dong@intel.com>

+++ hypervisor/arch/x86/vmexit.c
@@ -157,7 +157,7 @@ int vmexit_handler(struct vcpu *vcpu)
 	uint16_t basic_exit_reason;
 	int ret;
 
-	if ((int)get_cpu_id() != vcpu->pcpu_id) {
+	if (get_cpu_id() != vcpu->pcpu_id) {
 		pr_fatal("vcpu is not running on its pcpu!");
 		return -EINVAL;
 	}
